Create AppCalculator.vue
